+++ src/main/java/com/jcoronel/api/rest/barber/backend_barber/repositories/ServiceRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.jcoronel.api.rest.barber.backend_barber.entities.ServiceEntity;
 
-public interface ServiceRepository extends CrudRepository<ServiceEntity, Long>{
+public interface ServiceRepository extends CrudRepository<ServiceEntity, Integer>{
     
 }
